cambios de pago y orden de detalles y presupuesto de la orden

+++ database/migrations/2019_08_07_001927_add_value_to_currencies.php
@@ -14,7 +14,7 @@ class AddValueToCurrencies extends Migration
     public function up()
     {
         Schema::table('currencies', function (Blueprint $table) {
-            $table->float('value', 8, 8);
+            $table->decimal('value', 5, 4);
         });
     }
